finally got actual build working, move home back to the api route
