+++ src/components/Chat/Chat.tsx
@@ -23,7 +23,7 @@ import { throttle } from '@/utils/data/throttle'
 import { ChatBody, Conversation, Message } from '@/types/chat'
 import { Plugin } from '@/types/plugin'
 
-import HomeContext from '~/pages/home/home.context'
+import HomeContext from '~/pages/api/home/home.context'
 
 import Spinner from '../Spinner'
 import { ChatInput } from './ChatInput'
