+++ src/components/Chat/ChatInput.tsx
@@ -22,7 +22,7 @@ import { Message } from '@/types/chat'
 import { Plugin } from '@/types/plugin'
 import { Prompt } from '@/types/prompt'
 
-import HomeContext from '~/pages/home/home.context'
+import HomeContext from '~/pages/api/home/home.context'
 
 import { PluginSelect } from './PluginSelect'
 import { PromptList } from './PromptList'
