+++ src/components/Chat/ChatMessage.tsx
@@ -14,7 +14,7 @@ import { updateConversation } from '@/utils/app/conversation'
 
 import { Message } from '@/types/chat'
 
-import HomeContext from '~/pages/home/home.context'
+import HomeContext from '~/pages/api/home/home.context'
 
 import { CodeBlock } from '../Markdown/CodeBlock'
 import { MemoizedReactMarkdown } from '../Markdown/MemoizedReactMarkdown'
