+++ src/components/Chat/Temperature.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'next-i18next'
 
 import { DEFAULT_TEMPERATURE } from '@/utils/app/const'
 
-import HomeContext from '~/pages/home/home.context'
+import HomeContext from '~/pages/api/home/home.context'
 
 interface Props {
   label: string
