+++ src/components/Chatbar/Chatbar.tsx
@@ -14,7 +14,7 @@ import { LatestExportFormat, SupportedExportFormats } from '@/types/export'
 import { OpenAIModels } from '@/types/openai'
 import { PluginKey } from '@/types/plugin'
 
-import HomeContext from '~/pages/home/home.context'
+import HomeContext from '~/pages/api/home/home.context'
 
 import { ChatFolders } from './components/ChatFolders'
 import { ChatbarSettings } from './components/ChatbarSettings'
