+++ src/components/Chatbar/components/ChatFolders.tsx
@@ -2,7 +2,7 @@ import { useContext } from 'react'
 
 import { FolderInterface } from '@/types/folder'
 
-import HomeContext from '~/pages/home/home.context'
+import HomeContext from '~/pages/api/home/home.context'
 
 import Folder from '@/components/Folder'
