+++ src/components/Chatbar/components/Conversation.tsx
@@ -16,7 +16,7 @@ import {
 
 import { Conversation } from '@/types/chat'
 
-import HomeContext from '~/pages/home/home.context'
+import HomeContext from '~/pages/api/home/home.context'
 
 import SidebarActionButton from '@/components/Buttons/SidebarActionButton'
 import ChatbarContext from '@/components/Chatbar/Chatbar.context'
